backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sse-starlette==2.2.1
langgraph==0.2.60
langchain-anthropic>=0.3.0
yt-dlp==2024.12.13
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
